requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==3.1
django-cors-headers==3.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
future==0.18.2
idna==2.10
jsonschema==3.2.0
kiwisolver==1.2.0
kombu==5.0.0
matplotlib==3.3.0
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.1
Pillow==7.2.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
response==0.4.0
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
vine==1.3.0
pandas